[refactor] UI Layer에서 사용할 데이터를 담는 Model들을 Domain Layer로 이동시킴
