--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterBasic.kt
@@ -1,30 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import androidx.annotation.Keep
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@Keep
-@JsonClass(generateAdapter = true)
-data class CharacterBasic(
-    @Json(name = "character_name")
-    val characterName: String = "",
-
-    @Json(name = "world_name")
-    val worldName: String = "",
-
-    @Json(name = "character_gender")
-    val characterGender: String = "",
-
-    @Json(name = "character_class")
-    val characterClass: String = "",
-
-    @Json(name = "character_level")
-    val characterLevel: Int = 0,
-
-    @Json(name = "character_guild_name")
-    val characterGuildName: String? = null,
-
-    @Json(name = "character_image")
-    val characterImage: String = ""
-)
