--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterBasicResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterBasicResponse(
-    val status: ResponseStatus,
-    val characterBasic: CharacterBasic?
-)
