--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterDojangResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterDojangResponse(
-    val status: ResponseStatus,
-    val characterDojang: CharacterDojang?
-)
