--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterItemEquipmentResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterItemEquipmentResponse(
-    val status: ResponseStatus,
-    val characterItemEquipment: CharacterItemEquipment?
-)
