--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterOcidResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterOcidResponse(
-    val status: ResponseStatus,
-    val characterOcid: CharacterOcid?
-)
