--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterPopularity.kt
@@ -1,12 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import androidx.annotation.Keep
-import com.squareup.moshi.Json
-import com.squareup.moshi.JsonClass
-
-@Keep
-@JsonClass(generateAdapter = true)
-data class CharacterPopularity(
-    @Json(name = "popularity")
-    val characterPopularity: Int = 0
-)
