--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterPopularityResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterPopularityResponse(
-    val status: ResponseStatus,
-    val characterPopularity: CharacterPopularity?
-)
