--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterStatResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterStatResponse(
-    val status: ResponseStatus,
-    val characterStat: CharacterStat?
-)
