--- app/src/main/java/com/bodan/maplecalendar/data/dto/CharacterUnionResponse.kt
@@ -1,8 +1,0 @@
-package com.bodan.maplecalendar.data.dto
-
-import com.bodan.maplecalendar.data.ResponseStatus
-
-data class CharacterUnionResponse(
-    val status: ResponseStatus,
-    val characterUnion: CharacterUnion?
-)
