+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/CharacterDojang.kt
@@ -1,0 +1,8 @@
+package com.bodan.maplecalendar.domain.entity
+
+import androidx.annotation.Keep
+
+@Keep
+data class CharacterDojang(
+    val characterDojangBestFloor: Int = 0
+)
