+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/CharacterItemEquipment.kt
@@ -1,0 +1,9 @@
+package com.bodan.maplecalendar.domain.entity
+
+import androidx.annotation.Keep
+
+@Keep
+data class CharacterItemEquipment(
+    val itemEquipments: List<ItemEquipment>,
+    val itemTitle: ItemTitle?
+)
