+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/CharacterOcid.kt
@@ -1,0 +1,8 @@
+package com.bodan.maplecalendar.domain.entity
+
+import androidx.annotation.Keep
+
+@Keep
+data class CharacterOcid(
+    val ocid: String = ""
+)
