+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/CharacterPopularity.kt
@@ -1,0 +1,8 @@
+package com.bodan.maplecalendar.domain.entity
+
+import androidx.annotation.Keep
+
+@Keep
+data class CharacterPopularity(
+    val characterPopularity: Int = 0
+)
