+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/CharacterStat.kt
@@ -1,0 +1,8 @@
+package com.bodan.maplecalendar.domain.entity
+
+import androidx.annotation.Keep
+
+@Keep
+data class CharacterStat(
+    val finalStats: List<FinalStat>
+)
