+++ app/src/main/java/com/bodan/maplecalendar/domain/entity/ItemTotalOption.kt
@@ -1,4 +1,4 @@
-package com.bodan.maplecalendar.data.dto
+package com.bodan.maplecalendar.domain.entity
 
 import androidx.annotation.Keep
 import com.squareup.moshi.Json
